A new overcomplicated window recorder for every window.
This capturer is able to capture everything as long as it is in window mode and on top of every other window, since it cuts out the selected slice fromthe main screen.

+++ CriaAI-Core/Dependencies/BmpRenderer/Dependencies/libbmpread/bmpread.c
@@ -545,7 +545,7 @@ static int Validate(read_context * p_ctx)
             /* I believe C mandates that SIZE_MAX is at least 65535, so this
              * expression and the next are always safe.
              */
-            size_t colors = 1 << (size_t)p_ctx->info.bits;
+            size_t colors = ((size_t)1) << (size_t)p_ctx->info.bits;
             if(!(p_ctx->palette = (bmp_palette_entry *)
                  malloc(colors * BMP_PALETTE_ENTRY_SIZE))) break;
